requirements.txt
pyOpenSSL
Flask
flask-cors